# src/main/resources/static/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/usuario/{id}": {
      "get": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreateDto"
              }
            },
            "application/x-yml": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "usuario-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/criador/{id}": {
      "get": {
        "tags": [
          "criador-controller"
        ],
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "criador-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriadorCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/CriadorCreateDto"
              }
            },
            "application/x-yml": {
              "schema": {
                "$ref": "#/components/schemas/CriadorCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/CriadorResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "criador-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/avaliacoes/{id}": {
      "get": {
        "tags": [
          "avaliacao-controller"
        ],
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "avaliacao-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvaliacaoCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AvaliacaoCreateDto"
              }
            },
            "application/x-yml": {
              "schema": {
                "$ref": "#/components/schemas/AvaliacaoCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoCreateDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoCreateDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/AvaliacaoCreateDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "avaliacao-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/completar-cadastro/usuario": {
      "put": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "completarCadastroUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/completar-cadastro/criador": {
      "put": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "completarCadastroCriador",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriadorCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/assinaturas/{id}": {
      "get": {
        "tags": [
          "assinatura-controller"
        ],
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaResponseDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaResponseDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "assinatura-controller"
        ],
        "operationId": "update_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssinaturaCreateDto"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AssinaturaCreateDto"
              }
            },
            "application/x-yml": {
              "schema": {
                "$ref": "#/components/schemas/AssinaturaCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaCreateDto"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaCreateDto"
                }
              },
              "application/x-yml": {
                "schema": {
                  "$ref": "#/components/schemas/AssinaturaCreateDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assinatura-controller"
        ],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/auth/register/usuario": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "registerUsuario",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register/criador": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "registerCriador",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriadorCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssinaturaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idUsuario": {
            "type": "integer",
            "format": "int64"
          },
          "plano": {
            "type": "string",
            "enum": [
              "GRATUITO",
              "PREMIUM",
              "FAMILIA"
            ]
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "dataFim": {
            "type": "string",
            "format": "date"
          },
          "preco": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PlaylistDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idUsuario": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "visibilidade": {
            "type": "string",
            "enum": [
              "PUBLICO",
              "NAO_LISTADO",
              "PRIVADO"
            ]
          },
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistVideoDto"
            }
          }
        }
      },
      "PlaylistVideoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idVideo": {
            "type": "integer",
            "format": "int64"
          },
          "videoTitulo": {
            "type": "string"
          },
          "posicaoVideo": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UsuarioCreateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8,
            "pattern": "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
          },
          "interesses": {
            "type": "string"
          },
          "playlistDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistDto"
            }
          },
          "assinaturaDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssinaturaResponseDto"
            }
          }
        }
      },
      "UsuarioResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "interesses": {
            "type": "string"
          },
          "funcao": {
            "type": "string",
            "enum": [
              "USUARIO",
              "CRIADOR"
            ]
          }
        }
      },
      "CriadorCreateDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string"
          },
          "cpf": {
            "type": "string",
            "pattern": "\\d{11}"
          },
          "rg": {
            "type": "string",
            "pattern": "\\d{8,20}"
          },
          "senha": {
            "type": "string",
            "maxLength": 2147483647,
            "minLength": 8,
            "pattern": "^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d)(?=.*[@#$%^&+=!]).*$"
          },
          "formacao": {
            "type": "string"
          }
        }
      },
      "CriadorResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "formacao": {
            "type": "string"
          },
          "funcao": {
            "type": "string",
            "enum": [
              "USUARIO",
              "CRIADOR"
            ]
          }
        }
      },
      "AvaliacaoCreateDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "nota": {
            "type": "integer",
            "format": "int32",
            "maximum": 10,
            "minimum": 0
          },
          "comentario": {
            "type": "string"
          }
        },
        "required": [
          "comentario",
          "nota",
          "userId"
        ]
      },
      "AssinaturaCreateDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "plano": {
            "type": "string"
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "dataFim": {
            "type": "string",
            "format": "date"
          },
          "preco": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "dataInicio",
          "plano",
          "userId"
        ]
      },
      "AvaliacaoResponseDto": {
        "type": "object",
        "properties": {
          "nota": {
            "type": "integer",
            "format": "int32",
            "maximum": 10,
            "minimum": 0
          },
          "comentario": {
            "type": "string"
          }
        }
      },
      "AuthLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          }
        }
      }
    }
  }
}